Join functionality working for keyboard. Character select further development with render targets. Fixed healthbar issue and health component issue by deleting and then readding.
Input mapping working now finally had to set the input to ui and game in the character select so this might mess some other stuff up but can't test right now since I don't have a controller plugged in.

Character select rotatable player now actually has a texture and updated the player select blueprint a bit more to have a 2dscenerender in it but need to make sure that the rotatable players are the actual correct players which they probably aren't right now.
Also need to actually get the rotating part working.

Health component was broken for the health bars and setting attacking which was fixed by deleting the health component off the player blueprint and then adding it back again, IDK why this worked but yea.