"Leave" character select functionality setup. When clicking leave will reset the focus of any controllers that aren't the associated one so not stuck not focused on anything. Character select rendering done.
Players can now leave the character select once they have joined which will remove the character render and all the information.

Also when leaving the do once is reset for the join input action in the player select blueprint and also the focus for any player controllers that are on the leave button when it is pressed that aren't associated with the leave button itself have their focus set to their ready button.

The character renders now actually work properly with the correct materials shown and switching the character updating the mesh and weapons properly. Also had to add a new field for the character struct for the character's mesh.