Game instance and game modes update. Carrying over info from character select to the actual level. Janky fix for spawning players and their controllers.
All functionality for readying / unreadying and activating / deactivating players in the main menu game mode is now implemented being called from the character select widget.

Start game functionality is now working from the character select menu.

Carrying over the data from the game instance to the third person game mode is now working as intended except for one little mistake which is that 4 player controllers are spawned no matter how many players are actually active in the game but I've jankly fixed this by ignoring it and changing in the health bar widget to not use get player controller and instead use the get player character node instead. Setting the character info in the player blueprints is now done in the third person game mode.

Also getting the current level name and if it is equal to the dev level name then we are just using the test character info instead of the passed one and also all players are spawned since we are not going from the main menu and so can still easily test on the dev level without first going through the main menu.