Pausing functionality. Resume countdown functionality. Pause UI.
Pause input action implemented with being able to be done whilst the game is paused.

Pausing functionality implemented and putting the pause screen up and having all inputs and focus being right done for the pause screen.

New widget created for the resume countdown which just changes the text on construct and calls the animation 3 times with a finished event to change the text after each animation and then unpauses the game with the correct function in the thirdperson game mode.

Return button functionality done for pause menus since it was broken before.

Options and control buttons done for pause menu as well now.