Outlining text in clickable button widget. Ready button only shows when at least one player ready in character select menu. Increased rotation speed for rotatable render. Slight spacing and skew for text on clickable buttons. Shadow on clickable button text. Tweaked the character select screen. Tweaked outline and shadow and skew of various text in UI screens.
Lots of text in the UI now has outlines and shadows which make them look better in my opinion

The ready button in the character select screen is now only visible when at least one player is ready and will unready immediately as this is no longer true.

Changed the sizes of the character info in the character select screen to be better.

Made the spell info button text smaller

Also changed it so clicking the spell info button whilst the character info is up will change the character info button back to saying character info instead of character view.

Rotation speed of the rotatable player renders is now increased.