Boss trigger functionality done. Removed delay from thirdperson game mode. Payload decrease speed after taking damage for a set amount of time. Speed payload custom event to increase the speed of the payload. Start and stop payload triggers.
Boss trigger which only triggers once all active players have passed through it which will set focus boss on the camera to be active for a certain amount of time before setting it false again which returns the camera to focus back on the players. The focus time is a public exposed on spawn float variable that has a default value of 4.

Removed the slight delay from the begin play in the thirdperson game mode.
Also remember any time you want to get information from the game mode if it is also in the begin play do a tiny little delay before getting the info so it is correct and the game mode is done doing calculations etc.

2 new custom events in the payload - decrease speed and increase speed. New float variables: current speed and speed multiplier which act similarly to the movement manager component in that the current speed is finterp constantly using the speed multiplier and the base movement speed.

The 2 custom events pass whether the change is timed, the timed duration as well as the new speed multiplier.

Payload start moving trigger and payload stop moving trigger which basically just use the same functionality as the boss trigger but for setting the payload move speed multiplier to be 1 or 0 depending on the function.

When the payload takes damage it decreases its speed for an amount of time and amount of speed depending on the amount of damage it is taking - this could be tweaked to be the maximum health divided by the damage rather than right now with the damage being divided by arbitrary numbers.

Payload also completely stops moving now on its destruction.