Loading screen UI and functionality. Options screen resolution and volume functionality and UI done for now. Controls screen and functionality done.
Loading screen widget done both asthetically and functionally for now with the only changes needing to be made relating to the tips text that shows up using the text array in the loading screen widget that needs to be populated.

Options screen resolution and volume functionality and aesthetics done for now with the only additional functionality being the actual changing of the volume of the sound mix class overrides for the corresponding sound classes. Storing the volumes of the sound classes in the game instance so can keep track of them not in the clickable button widget.

Controls screen functionality and aesthetics done for now if needing to update the actual controls in the future then just change the controls widgets for all keyboard and mouse, playstation and xbox. Using Xelu's free controller and key prompts for the visuals of the controls.

Started working on the credits screen with the heading buttons functionality partially implemented and the aesthetics partially started. Want to commit this right now so I have a commit for today tbh. Gonna finish the credits screen before sleeping though for the most part.