Funky noise UI material made with help from tutorial and put in background of main menu.
https://www.youtube.com/watch?v=z06wAYHNXxQ
Used this tutorial to make the noise material and then messed with it a lot to change how it looked and then put it as the image to draw for the background colour of the main menu. Probably not going to put it on other screens since it is a bit distracting and can take away from being able to read the text and such so prob gonna stay as a main menu only thing. Can always change it in the future anyway.