Carriage Mesh put in. Carriage moving characters and physics bodies out of the way. Carriage moving wheels with speed.
Carriage mesh from the wild west vol.2 free asset pack from Unreal asset store put into the game. https://www.fab.com/listings/e9700286-334e-4230-9de4-2d8db738fdeb

Carriage wheels are different static meshes in the payload blueprint. Coffin meshes from the same asset pack put in the inside of the payload's carriage mesh.

When sweeping on set actor location for the payload we now check if blocking hit and if so we switch on the hit component's object type. If it is a physics body then we add a force to it depending on the speed of the carriage. If it is a player or enemy then we add an input vector dependent on the speed of the payload.

The carriage wheels are constantly rotating with the amount to add for the local rotation pitch being dependent on the speed of the payload.