Fixed an issue where the players weren't in game only input when loaded into normal level. Character select update.
Now in the thirdperson game mode we set the input mode to game only after adding the player to the players array so that the input works as intended for players in the actual playable level.

Character select update: when leaving it now also sets the player that left to no longer be ready. Also once ready is set cannot change the current character until the player unreadys. Also changed the padding of the character info so that it all fits on and there is space left for if there is more info than expected I think.