Payload update. Fixed aoe persistence collision. Character Select join and leave widget animation. Changed colour scheme to green for UI. Started working on the payload spline system but not complete.
Payload active bool to determine if we can actually "slow" the payload so that when it is not active now it should not return to moving after being damaged and also the tick will turn off when no longer active and will be enabled again when set active again.

Aoe persistence collision is now updated properly since it wasn't after adding in new collision responses.

Lots of UI changes specifically in the character select menu. Changed colour scheme to be green oriented dark.

Character select player join and leave now have animations tied to the rotatable render so more juice now.