Spline system fully working now for the payload. Added collision response for payload.
Realised that the issue with the spline system for the payload was due to the camera border blocking the payload. Added a new collision object response for the payload specifically and updated everything to use this properly.

Possibly fixed the melee weapon traces to trace for players, enemies, and the payload instead of just pawns and physics bodies.