Put ultra dynamic sky and altered the leaves colouring and intensity.
Ultra dynamic sky and weather system implemented (put into the project) and set it up in the map level for night time thunderstorm, cloudy and foggy for now.

Altered the leaves of the trees material instance so that they are more of a dark green to match the artists colour boards / palette and also made the emissive strength a bit lower so they don't glow as much in the dark.