Added in From Main Menu Bool to game instance.
New bool variable in the game instance which will determine whether we load in players for all the player starts or just the ones from the character select. This is default as false and then only set false when starting the game from the character select screen. This means that it is now easier to just load into a level with a set number of players without having to go through the main menu and character select process but this should still work for the final build of the game going through the menus normally and stuff.