Removed payload health healing print string. Fixed some stuff with enemy AI. Showcase enemy rows in enemy data table.. Movement manager small change. Remove HUD with T will now add HUD back if clicking T whilst HUD not visible. New behaviour tree for evading for ranged enemies. Camera track players and camera track payload triggers.
Enemy can ranged fixed I think with the bools.

Made some showcase enemy rows in enemies data table.

Setup videos in dev level.

Made 2 new triggers for tracking players and tracking payload with camera.

New behaviour tree for ranged enemies evading when they cannot do a ranged attack which just does a eqs and moves to it so that they reposition if the player gets to close to a ranged enemy and so they don't just stand there until they are in range to attack again.