Enemy AI perception update.
On Target perception event used now for adding to new actor array of perceived targets and then if the stimulus is not successfully sensed it is removed from the array.

On target perception Forgotten we remove the actor from the perceived targets array and then if the array is empty then after 2 seconds the enemy returns to being passive / idle state and removes the focus.

The update state event now only sets the state if the new state is not the same as current state.

The on target perception updated event now loops through the perceived targets array if the stimulus is successfully sensed and then it does the rest of the state machine code for each loop.

Also got rid of the cosmetic errors for the state machine by checking if the health v2 component is valid before doing the checks with functionality not changing but errors being removed.