Fixed aim assist. Fixed a small issue with changing the capsule half height for being downed and revived. Fixed finterp speed being too slow. Aim assist scale now determined by the spell's aim.
Aim assist was actually broken if the array was empty which is fixed now.

The rotation widget in the player blueprint is now properly moving when the player hit box changes for being downed and revived.

Movement manager finterp speed is now way way faster because it was incredibly slow so the changes to speed were negligible but this is now changed.

Aim assist size is now dependent on the spell general info so dependent on the spell itself.