Enemy AI perception update.
Enemies will now automatically update their chasing target and focus when a player enters a collision around the enemy.

Also when an enemy dies then a multi-sphere trace is done for other enemies in a certain range to then alert the enemies and change their chasing character and focus target on the character that has killed the enemy. Also had to compile the ranged attack component to then fix the spell firing and spell info thing after changing the name of the spell info struct and then renaming it back which didn't fix it.