Enemy change spawn rate and amount trigger created and implemented. Fixed enemy spawning director spawning amount for locations.
New blueprint for a trigger which will increase the spawn rate of the enemy spawning director and also increase the spawn amount.

Also fixed the enemy spawning director spawn amount location so that the group of enemies don't all spawn in the exact same location and so that the location is re-calculated everytime we loop through for the amount of enemies to spawn.