Spell casting from hands as well as from a weapon now.
New bool variable in spell general which determines if the spell will be shot from the character's ranged weapon or from their right hand (for now could be changed in the future to be a different socket or something).

2 New sockets in the mannequin skeleton for the right hand cast location and another holster weapon socket on the back but with a different rotation and location than the original holster weapon socket.

New enum for the spell cast method with 2 enumerators: weapon and hand.

New bool in the attach weapon component for if the ranged weapon is holstered whilst casting. New check in the switch weapon function after checking if the melee holstered bool is true and the switch to melee bool is false which checks if the ranged holstered for casting bool is true which will then unholster the ranged weapon and holster the melee weapon and also now set the ranged holstered whilst casting to be false.

New select for find look at rotation and a new select for the location of the spell to be shot from in the ranged attack component shoot function. Will select the ranged weapon mesh's socket location if the spell cast method in the spell is equal to weapon and if it isn't weapon then it selects the vector from the get hand cast loc function in the bpi has ranged attack. This function gets the socket location of the character mesh with the socket name "Hand_Cast_Loc_R".

Some changes in the attack handler for switching the character holstering depending on if the spell is casting based on hand or weapon.