Changed enemy spawning change trigger and by proxy changed the update spawn rate and spawn amount functions in spawning director.
The enemy change trigger is now called the enemy spawning change trigger to be more accurate to its function.

The 2 variables in the enemy spawning change trigger have been changed from being increases to their respective variables in the enemy spawning director to be the actual new values of these respective variables so it is easier to make different waves spawning without having to do calculations.

The 2 functions for updating these values have been changed in the enemy spawning director component to be setting them instead of increasing / decreasing the values and clamping them.