Tweaking cooldowns for EQS runs in trees. Also possibly fixed a very very small issue with the chasing payload melee tree's attacking task. Made some new folders in the SFX folder. UI sounds.
Longer cooldowns for the EQS in the enemy's behaviour trees so that fewer EQS tasks are run constantly and I think a 0.8 second is appropriate for reducing the amount of tasks whilst keeping the performance of the enemy's pathfinding working as intended.

Also noticed in the chasing payload melee behaviour tree one of the passed parameters for the attack task was not set properly so set this, this could be a recurring thing with behaviour trees as experienced before not sure yet.

Made some more folders in the SFX folder for future audio assets.

Got 2 sounds from JSFXR: button click sound, and button hover sound and created sound cues for these with modulators to change the pitch so that they are not the same every time they are played. Added the play sound 2D nodes for these in the clickable button widget and set the volume for the hovered sound really low since it happens a lot of the time.