Camera number tweaks. Holding next or previous spell cycles through now. Debug restart level button for showcase if needed.
Tweaked the angle of the camera a little.

Changed the next and previous spell input actions to also have hold triggers and also changed the functionality associated with this input action so that holding down either of the triggers will now cycle through in the direction of the input.

Debug restart level button in third person character blueprint under - button on numpad which just gets the current level name and then opens the level by name of this to restart the level. This is just a backup for the showcase at the end of semester incase we need it.