AAAAAAAAAAH OMG THIS IS DRIVING ME INSANE. Charmed / turned status effect. Also fixed an issue with clearing timers for status effects. Fixed an issue with melee enemies and the evading tree.
Majority of the charmed / turned status effect is in and working to some extent just some things driving me insane so im doing it tomorrow or whenever.

Also fixed some stuff with the clear status effect timer in the status effect timer component since it was only ever clearing the bleeding timer since none of the other timers were properly being cleared etc so using a switch and yea.

Also fixed a small issue with the melee enemies and their use of the evading behaviour tree and tasks inside it so the can ranged decorator now checks if they actually are a ranged enemy and if so they cannot go into the evade tree so yea.