Charmed / Turned Speed change and updating target whilst charmed / turned with new field in the charmed / turned status effect struct.
Changing the speed of the charmed/ turned target using set speed multiplier in movement manager component using reference to owner of the status effect handler component and the percentage move speed slow field in the charmed / turned status effect struct.

Also altered the interp speed again in the movement manager component since it was still too slow in my opinion.

New field in the charmed / turned status effect struct for how many times during the duration of the effect should the traces be done to then update the target of the charmed / turned target. This is then done in the charmed / turned status effect function in the status effect handler component using game time in seconds and this new field as well as the duration field to check if the time has surpassed the last time checked plus the duration / the count for how many times it should be checked. If the times to check if 0 then it will do the initial one and then never do another check until it is reapplied etc.