Charmed / turned optimisation and commenting. Can no longer switch spells whilst charging a spell. Spell switching properly fixed. Attracting fixes. Juice spell HUD for swapping. Status Effect HUD representation
Changed some order of operations in the turned / charmed function to help optimise it a little bit and also commented it.

Also commented some other stuff like the attracting component for spells and some of the spell projectile blueprint.

Spent some time messing around with the attracting component and think its properly right now.

Can no longer switch spells whilst charging up a spell.
Tweaked the input actions for the next and previous spell swapping keys and also added in some more stuff to remove the glitch where clicking next or previous would skip 1 spell because of the hold functionality so all of this is now correct and functionality I'm pretty sure.

New animation for switching spells which just adds some more juice to it.

Did status effect representation in the HUD haven't done this properly for the enemies in the world but should be decently easy since it will be using similar functionality and probably the same widgets just with it being in world space instead of on the HUD. But the representations work with the status effect being reapplied and they have juice as well just need the actual icons for the status effects.

Finally also changed the fire status effect in the enum status effects to be burning since fire was just stupid and not in line with the names of the other status effects.