Fixed various things. Issue with spell fire location for enemies (stupid mistake). New get spells interface function for characters which kinda fixed the spells component (may have to remake spells component like the health component). Setup struct for spell VFX. Fixed status effect background. Enemy spawning director further optimising. Started creating character-specific spell names and setting up in data table.
Enemy get hand fire loc function wasn't setup so they weren't shooting properly.

Did some extra checks for spell cooldowns to get rid of error messages.

New function for getting spell names for the spells component instead of getting the character info and then getting the spell names which may have fixed it idk. Something changed that fixed the spell component it may be that the spell component is corrupted or something like the health component was and may have to remake it at some point.

Setup a struct for the spell VFX including: shoot vfx, charging vfx, end of life vfx and travelling vfx which currently are niagara systems but if using emitters may have to change this. Will actually setup the use of these vfx later.

Status effect background is now transparent.

Enemy spawning director further optimised to not have the tick enabled in unneccessary circumstances so when the spawn rate is 0 or the spawning amount map is empty then the tick is not enabled.

Started creating character-specific spell names in the enum spell names and the non-custom spells data table.