Camera shake for shooting spell changed, camera shake for spell EOL, camera shake for AOE blast, camera shake for entering aoe persistence and leaving aoe persistence. Slightly altered aiming start point.
Now using the distance from the camera for the spell shoot camera shake. New camera shake using the same shoot camera shake blueprint for the spell EOL, similarly for the AOE blast, entering the AOE persistence and leaving the AOE persistence, and also did a little camera shake for the piercing characteristic piercing something.

For piercing to work for the objects the object needs to have generate overlap events activated.

Aiming point for spells traces now comes from the actor eyes point instead of the pivot