Controller rotation update and bettering. Mouse rotation new setup which is better, would advance more if had time but need to focus on other stuff now.
New input action specifically for looking with controller so can have separate functionality.
Controller look breaks the 2d vector for the action value which is the value of the right thumbstick and then makes a new vector with the y inverted and set as the x and the x value of the input action as the y value of the vector and then we make rotation from axes from this vector with it being the forward axes and then set that as the control rotation of the current player.

For the mouse looking we now use the get hit result under cursor by channel visibility and then set the control rotation from a find look at rotation from the player's location to this vector either being the location of the hit or the trace end depending on if there was a blocking hit and then when setting the control rotation we only use the z axis.