Enemies will now choose a random spell from their spells array in the enemy data. Ida and kamarra rigged skeletal meshes now in game. Using ida as the original skeleton to have everyone else be compatible with hopefully this works. Also put jess' model in rigged. Added FXVariety Pack, MixedVFX, and Rocket Thruster Exhaust FX to the project. Charlie mesh in (bit janky)
New functionality within the spells component to check if the owner is an enemy and if so then we select a random spell from their array of spells to choose from and use which will help to reduce the time I need to spend creating specific enemy rows in the enemy data table and also will create more uniqueness from the enemies spawning since they will use different spells each time.

Skeletal meshes of low poly characters in the game now just need to buy the plugin and do all the animations and make these skeletal meshes to have root motion available and then also fix any inevitable errors and glitches that happen.

Added 3 asset packs of VFX to the project.