Mouse Controls update for mouse cursor. Fixed special melee spells not working without having to attack first. Also added a check in the attack handler for can attack being not melee attack in the ranged attack component.
Mouse cursor being set visible and also the input game mode being set to be game and UI when moving the mouse constantly so that the mouse doesn't disappear and so that you can still rotate with the mouse whilst using the mouse buttons. Retriggerable delay to set mouse cursor not visible and input game only, delay only has a 0.05 second delay.

Mouse rotation still works with other characters using controller as well.

Fixed the special melee spells not working for the first time when trying to use them which was because the name of the weapon being used wasn't being gotten so fixed that in the animation blueprint adding another check and getting the melee weapon name. Also added a check in the attack handler for checking if melee attacking in the ranged component so not spamming it whilst its being done.