Bought mixamo converter plugin AND IT WORKS. Also fixed an issue with spell hand casting.
Added the mixamo converter plugin and enabled it.

This is the process for getting the animations to work with the mixamo skeletons / how we / I am going to do it:

Create an IK retargeter for the skeleton that is compatible with all other skeletons (ida's skeleton).

Open the IK retargeter called RTG_Mannequin_<MixamoSkeletonName>
Select all animations wanting to retarget from UE to mixamo and then boom done.

For root motion to work with the mixamo skeletons you need to right click on the skeleton and click retarget mixamo skeleton but since they are all compatible skeletons with the ida skeleton we shouldn't need to keep retargeting animations after doing it with the first skeleton.

For Ik to work for a skeletal mesh that you are putting in make to make a IK retargeter since you actually don't need to use it to retarget but instead need it for the IK rig to be used in the animation blueprint.

I have retargeted all the animations that are being used in the animation blueprint and copied the original animation blueprint to be a new one for mixamo and setup all the animations in there properly as well as the layered blend per bones since they will be different names. Might need to redo emmas character since its a bit janky and maybe jess' as well.

Tried changing the control rig bones to be something different for the IK but I don't really know if it changed anything.