Enemy spawning director update. Boss second phase implemented (not tested). Slightly altered attach weapon attachment rules.
Enemy spawning director now uses get owner instead of just getting the payload for positions for spawning so that it works for spawning enemies for the boss second phase. Cannot remove the variables without the engine crashing though for some reason so leave them in for now.

New field in the enemy struct for the percentage of health the boss needs to be at to start the second phase. In the health component tack damage function after checking if the owner implements the boss interface we then get the enemy info and see if the current health is less than or equal to the maximum health * percentage health to start second phase. If so we call the activate 2nd phase function in the boss enemy.

This new function in the boss enemy has a do once to then update its own enemy spawning director with some enemies to spawn and a new spawn rate. This should work, granted I've not tested it yet.

Also added in the sockets for jess' character but will need to update the weapon meshes for her character since its the boss if it has weapons, since it could just use hand casting and swipes with its hands, need to be careful since its going to add more work though if we have unhanded attacks so might have to use some sort of weapon.