Changed AOE persistence adding status effects. Spell VFX functionality implemented. AOE persistence Colour setup.
AOE persistence when adding status effects on event tick will only add them if the target doesn't have the status effect gameplay tag. But the event tick interval has now been decreased to 0.5.

Added functionality for the use of niagara systems with the spells including: shooting VFX, charging VFX, spell End of life VFX, and finally spell travelling VFX. Took way longer than it should have since I was just being stupid and not testing it properly so yea.

Colour of the AOE persistence is now set in the spell AOE info.

New field in the spell general info struct for how long it should take for the collision to be renabled for the spell after being shot which should be really low or 0 but depending on the scale it might need to be a bit higher like 0.5 seconds.

The thruster VFX look really cool for the spell travelling effects.