Charging VFX Change. Split VFX.
Charge VFX now only gets destroyed / disabled when actually shooting and not when stopped charging for completing the charge (obviously still gets destroyed / disabled if stopped charging and not charged long enough to actually cast).

Split VFX so when a split spell splits then it also does the shoot VFX so there's some indication.

Changed the duration of the blood decal and the start delay so that the blood puddles last way longer.