Switch Spell Task for enemies. Boss Trigger Update. Spells Data table. Player info data table. Enemies Data table.
New task for enemies to do which switches the current spell by just calling the next or previous spell from the spells component of the owner of the task which will only be done if the owner is not already charging a spell. Also it is a chance between switching to the next spell or the previous spell. (Not tested but should theoretically work)

Done about 24 spells in the spells data table whilst also keeping the originals but they probably won't be used.

Changed some stuff in the enemies data table, player info data table, character names enum, and enemies enum but still need to properly populate the data for the player info and the boss info so will do that after this commit. Just wanted to commit this so can roll back without losing everything.