Fixed character select max enumerator character glitch. Sound classes and sound mixes added and setting their volumes in game instance.
The character select names were bugged to say enum character max for the rotatable renders which has been fixed.

Sound classes created for the 4 different types of audio and sound mixes for the 4 types as well.

Functionality added for changing the volume of the sound mix class overrides to actually change the sounds volume across all the sound cues. Also changed all pre-existing sounds to be using the correct sound classes.