Controller rotation fixed to work based on the rotation of the camera now as well. Removed underscores for the spell names in the character info widget. Inescapable love tweak. Lowered all spell charge times. Boss enemies aim assist scale increased by 3 times for their spells in the spells component. Fixed enemy charging spells. Think fixed the charging spell speed change in ranged attack component. Made deaths rip particle much bigger. Changed the pivot of the arrow mesh so its the correct way for the particle effects using the arrows.
For the controller look rotation we now combine the camera rotation with the intended rotation of the player to then get the correct rotation based on the camera's rotation.

Using the code for removing the underscores from the spell names in the spell / ability widget in the character info widget to get rid of the underscores for the spells / abilities there as well now.

Inescapable love now has longer time before collision active so not hitting the character casting.

Trying to fix the boss enemy and the enemies charging spells being interrupted by starting the spell again.

Attack task was breaking for getting the current spell since it was using the attack handler component instead of the controlled pawn so fixed that.

Might need to work more on the boss itself and test it without the payload in its view.

Changed some of the other spell info.