Fixed bordering issue for health bar. Did same thing as normal health bar for carriage health bar and boss health bar. New repeating tiling material for health bar / UI blood.
The health bar no longer has the middle border thing so just one straight thing
Also updated the animation to use the overlay
Changed the carriage and the boss health bars to be the same as the player health bars.
Created a material that pans noise for the blood so its a bit more interesting.
Changed the scale of some of the things in the health bars.
Made the health bars in the HUD a bit bigger.