Animation blueprint optimisation. Increased spawn time for all enemy spawn triggers in level. Commented some stuff. Removed some unused code.
Animation blueprint has now been optimised to use multithreading / using thread safe update animation instead of the update animation event since that can't be multithreaded.
Had to remove everything that was done in the event graph and instead use a different method that is thread safe.
It should be more efficient and better performance but can't really tell right now.

Changed the spawn rates of all enemy spawn triggers in the level to be longer so that enemies don't overwhelm so quickly.

Commented some stuff and removed some unused code from various blueprints.