Fixed the AI controller.
The enemy AI controller was broken which was why the enemies movement was working properly. Basically to fix this you have to change the parent class of the controller to something random and then compile and then change the parent class back to the ai controller class so if this happens again in the future just do this.

THIS TOOK SO UNNECESSARILY LONG TO FIND OUT LITERALLY TOOK ME HOURS OF REMAKING THE ENTIRE ENEMY AND ENEMY AI TO FIND IT WAS THE CONTROLLER THAT WAS THE PROBLEM YAYAYAYYAYAYAYAYAYAYAY.