Enemy spawning director change and fix. Modular difficulty based on number of players in the game. Changed some spell infos.
Changed the enemy spawning director to use a function with timer instead of using the event tick which is now using the spawn rate properly.
This should have fixed the issue with the timing for the spawning not being correct.

Modular difficulty based on how many enemies will spawn in the enemy spawning director depending on how many players are in the game.

Eldritch blast and the bombs have different info and solar flare.